Split log scanner tests by scope
