+++ Tests/OpenUsageTests/CodexLogUsageScannerTests.swift
@@ -1,13 +1,8 @@
 import XCTest
 @testable import OpenUsage
 
-/// `CodexLogUsageScanner` against fixture rollout files — parsing, totals deltas, model tracking,
-/// subagent replay skipping, dedup, aggregation, and the end-to-end scan. Fixture semantics ported
-/// from ccusage's Codex adapter tests.
-final class CodexLogUsageScannerTests: XCTestCase {
-    private let pricing = TestPricing.bundled
-
-    private func fixedRates(_ input: Double = 1000, _ output: Double = 3000) -> ModelPricing {
+enum CodexLogUsageScannerTestFixtures {
+    static func fixedRates(_ input: Double = 1000, _ output: Double = 3000) -> ModelPricing {
         ModelPricing(
             supplement: PricingSupplement(),
             primary: PricingCatalog(entries: ["gpt-5.2": ModelRates(
@@ -17,7 +12,12 @@ final class CodexLogUsageScannerTests: XCTestCase {
             secondary: PricingCatalog(entries: [:])
         )
     }
+}
 
+/// `CodexLogUsageScanner` against fixture rollout files — parsing, totals deltas, model tracking,
+/// subagent replay skipping, dedup, and aggregation. Fixture semantics are ported
+/// from ccusage's Codex adapter tests.
+final class CodexLogUsageScannerTests: XCTestCase {
     // MARK: - Line parsing
 
     func testLastTokenUsageWinsOverTotalsDelta() {
@@ -274,7 +274,7 @@ final class CodexLogUsageScannerTests: XCTestCase {
                 makeEvent("2026-05-12T09:00:00.000Z"),
                 makeEvent("2026-05-13T08:00:00.000Z")
             ],
-            since: .distantPast, pricing: fixedRates(), fastTier: false
+            since: .distantPast, pricing: CodexLogUsageScannerTestFixtures.fixedRates(), fastTier: false
         )
 
         XCTAssertEqual(scan.series.daily.count, 2)
@@ -299,7 +299,7 @@ final class CodexLogUsageScannerTests: XCTestCase {
             total: 150
         )
         let scan = CodexLogUsageScanner.aggregate(
-            events: [event], since: .distantPast, pricing: fixedRates(), fastTier: false
+            events: [event], since: .distantPast, pricing: CodexLogUsageScannerTestFixtures.fixedRates(), fastTier: false
         )
 
         var lines: [MetricLine] = []
@@ -323,7 +323,7 @@ final class CodexLogUsageScannerTests: XCTestCase {
         // The same event parsed from a copied session file counts once.
         let event = makeEvent("2026-05-12T08:00:00.000Z")
         let scan = CodexLogUsageScanner.aggregate(
-            events: [event, event], since: .distantPast, pricing: fixedRates(), fastTier: false
+            events: [event, event], since: .distantPast, pricing: CodexLogUsageScannerTestFixtures.fixedRates(), fastTier: false
         )
 
         XCTAssertEqual(scan.series.daily.first?.totalTokens, 150)
@@ -332,7 +332,7 @@ final class CodexLogUsageScannerTests: XCTestCase {
     func testAggregateCachedTokensPriceAtCacheReadRate() {
         let scan = CodexLogUsageScanner.aggregate(
             events: [makeEvent("2026-05-12T08:00:00.000Z", input: 1000, cached: 400, output: 0)],
-            since: .distantPast, pricing: fixedRates(), fastTier: false
+            since: .distantPast, pricing: CodexLogUsageScannerTestFixtures.fixedRates(), fastTier: false
         )
 
         // 600 non-cached x $1000/M + 400 cached x $100/M = 0.6 + 0.04.
@@ -342,11 +342,11 @@ final class CodexLogUsageScannerTests: XCTestCase {
     func testAggregateFastTierDoublesWhenNoExplicitMultiplier() {
         let base = CodexLogUsageScanner.aggregate(
             events: [makeEvent("2026-05-12T08:00:00.000Z")],
-            since: .distantPast, pricing: fixedRates(), fastTier: false
+            since: .distantPast, pricing: CodexLogUsageScannerTestFixtures.fixedRates(), fastTier: false
         )
         let fast = CodexLogUsageScanner.aggregate(
             events: [makeEvent("2026-05-12T08:00:00.000Z")],
-            since: .distantPast, pricing: fixedRates(), fastTier: true
+            since: .distantPast, pricing: CodexLogUsageScannerTestFixtures.fixedRates(), fastTier: true
         )
 
         XCTAssertEqual(
@@ -367,7 +367,7 @@ final class CodexLogUsageScannerTests: XCTestCase {
                 makeEvent("2026-05-12T08:00:00.000Z", model: "mystery-model-9"),
                 makeEvent("2026-05-12T09:00:00.000Z")
             ],
-            since: .distantPast, pricing: fixedRates(), fastTier: false
+            since: .distantPast, pricing: CodexLogUsageScannerTestFixtures.fixedRates(), fastTier: false
         )
 
         // Unpriceable tokens never enter the displayed totals — they surface only through the
@@ -381,7 +381,7 @@ final class CodexLogUsageScannerTests: XCTestCase {
     func testAggregateUnknownModelOnlyLeavesDayUnbacked() {
         let scan = CodexLogUsageScanner.aggregate(
             events: [makeEvent("2026-05-12T08:00:00.000Z", model: "mystery-model-9")],
-            since: .distantPast, pricing: fixedRates(), fastTier: false
+            since: .distantPast, pricing: CodexLogUsageScannerTestFixtures.fixedRates(), fastTier: false
         )
 
         // A day with nothing priceable produces no series entry at all (→ "No data"), but the
@@ -395,109 +395,10 @@ final class CodexLogUsageScannerTests: XCTestCase {
         let scan = CodexLogUsageScanner.aggregate(
             events: [makeEvent("2026-01-01T08:00:00.000Z"), makeEvent("2026-05-12T08:00:00.000Z")],
             since: OpenUsageISO8601.date(from: "2026-05-01T00:00:00.000Z")!,
-            pricing: fixedRates(), fastTier: false
+            pricing: CodexLogUsageScannerTestFixtures.fixedRates(), fastTier: false
         )
 
         XCTAssertEqual(scan.series.daily.map(\.date), ["2026-05-12"])
     }
 
-    // MARK: - End-to-end scan
-
-    func testScanReadsSessionsAndArchivedSessions() async throws {
-        let day = ISO8601DateFormatter().string(from: Date().addingTimeInterval(-3600))
-        let home = try CodexLogFixture.makeHome(files: [
-            "sessions/2026/05/rollout-a.jsonl": [
-                CodexLogFixture.turnContext(timestamp: day, model: "gpt-5.2"),
-                CodexLogFixture.tokenCount(timestamp: day, last: CodexLogFixture.usage(input: 100, output: 50))
-            ].joined(separator: "\n"),
-            "archived_sessions/rollout-b.jsonl": CodexLogFixture.tokenCount(
-                timestamp: day, last: CodexLogFixture.usage(input: 30, output: 20), model: "gpt-5.2"
-            )
-        ])
-        let scanner = CodexLogFixture.scanner(home: home)
-
-        let scan = await scanner.scan(pricing: fixedRates())
-
-        XCTAssertEqual(scan?.series.daily.reduce(0) { $0 + $1.totalTokens }, 200)
-    }
-
-    func testScanPrefersActiveSessionsCopyOverArchivedDuplicate() async throws {
-        let day = ISO8601DateFormatter().string(from: Date().addingTimeInterval(-3600))
-        let content = CodexLogFixture.tokenCount(
-            timestamp: day, last: CodexLogFixture.usage(input: 100, output: 50), model: "gpt-5.2"
-        )
-        let home = try CodexLogFixture.makeHome(files: [
-            "sessions/rollout-a.jsonl": content,
-            "archived_sessions/rollout-a.jsonl": content
-        ])
-        let scanner = CodexLogFixture.scanner(home: home)
-
-        let scan = await scanner.scan(pricing: fixedRates())
-
-        // Same relative path in both dirs = the same session archived; identical events dedupe
-        // anyway, but the discovery-level rule keeps it to one parse.
-        XCTAssertEqual(scan?.series.daily.reduce(0) { $0 + $1.totalTokens }, 150)
-    }
-
-    func testScanReturnsNilWithoutCodexHome() async {
-        let scanner = CodexLogFixture.scanner(home: nil)
-        let scan = await scanner.scan(pricing: fixedRates())
-        XCTAssertNil(scan)
-    }
-
-    func testScanCachesUnchangedFilesAndPicksUpNewOnes() async throws {
-        let day = ISO8601DateFormatter().string(from: Date().addingTimeInterval(-3600))
-        let home = try CodexLogFixture.makeHome(files: [
-            "sessions/rollout-a.jsonl": CodexLogFixture.tokenCount(
-                timestamp: day, last: CodexLogFixture.usage(input: 100, output: 50), model: "gpt-5.2"
-            )
-        ])
-        let scanner = CodexLogFixture.scanner(home: home)
-
-        let first = await scanner.scan(pricing: fixedRates())
-        XCTAssertEqual(first?.series.daily.reduce(0) { $0 + $1.totalTokens }, 150)
-
-        try CodexLogFixture.tokenCount(
-            timestamp: day, last: CodexLogFixture.usage(input: 30, output: 20), model: "gpt-5.2"
-        ).write(to: home.appendingPathComponent("sessions/rollout-b.jsonl"), atomically: true, encoding: .utf8)
-
-        let second = await scanner.scan(pricing: fixedRates())
-        XCTAssertEqual(second?.series.daily.reduce(0) { $0 + $1.totalTokens }, 200)
-    }
-
-    /// Manual parity harness against the real logs on this machine: prints per-day totals to compare
-    /// with `ccusage codex daily --json --offline`. Gated like the other live tests.
-    func testParityAgainstRealLocalLogs() async throws {
-        try XCTSkipUnless(ProcessInfo.processInfo.environment["OPENUSAGE_CODEX_PARITY"] == "1")
-        let scanner = CodexLogUsageScanner()
-        let result = await scanner.scan(pricing: TestPricing.bundled)
-        let scan = try XCTUnwrap(result)
-        for day in scan.series.daily.sorted(by: { $0.date < $1.date }) {
-            print("PARITY \(day.date) tokens=\(day.totalTokens) cost=\(day.costUSD.map { String(format: "%.4f", $0) } ?? "nil")")
-        }
-        if !scan.unknownModelsByDay.isEmpty {
-            print("PARITY unknown models: \(scan.unknownModelsByDay)")
-        }
-    }
-
-    func testScanPricesRealCodexModelsFromBundledSnapshots() async throws {
-        let day = ISO8601DateFormatter().string(from: Date().addingTimeInterval(-3600))
-        let home = try CodexLogFixture.makeHome(files: [
-            "sessions/rollout-a.jsonl": [
-                CodexLogFixture.turnContext(timestamp: day, model: "gpt-5.3-codex"),
-                CodexLogFixture.tokenCount(
-                    timestamp: day,
-                    last: CodexLogFixture.usage(input: 1_000_000, output: 0)
-                )
-            ].joined(separator: "\n")
-        ])
-        let scanner = CodexLogFixture.scanner(home: home)
-
-        let scan = await scanner.scan(pricing: pricing)
-        let today = scan?.series.daily.first
-
-        // gpt-5.3-codex must resolve in the bundled LiteLLM snapshot and price > $0.
-        XCTAssertTrue(scan?.unknownModelsByDay.isEmpty ?? false)
-        XCTAssertGreaterThan(today?.costUSD ?? 0, 0)
-    }
 }
